requirements.txt
llama-cpp-python
psutil
pyyaml
python-dotenv
